perf: rust端不再操作config
